requirements.txt
fastapi
uvicorn
pika              
httpx
python-jose       
passlib[bcrypt]   
requests
sqlalchemy
pydantic
pydantic-settings